Extract reusable backend app construction
